Added Autogened code to repo
This way, it’s a LOT easier to clone and build, but still not binary.
This is how the super cool C# steamkit guys do it, so I’ll gie it a go.
I still use my shell script to generate these.

Note: MessagesClientserver will probably not look good in an ide, ie
IntelliJ barfs with it, it’s just too big.